refactor(ramps): move getProviderWebviewColors to utils (TMCU-1087)
