+++ app/components/UI/Ramp/Views/Checkout/Checkout.tsx
@@ -46,7 +46,7 @@ import { useStyles } from '../../../../hooks/useStyles';
 import styleSheet from './Checkout.styles';
 import { useTheme } from '../../../../../util/theme';
 import { AppThemeKey } from '../../../../../util/theme/models';
-import { getProviderWebviewColors } from '../../constants/providerWebviewColors';
+import { getProviderWebviewColors } from '../../utils/getProviderWebviewColors';
 import Device from '../../../../../util/device';
 import { shouldStartLoadWithRequest } from '../../../../../util/browser';
 import { CHECKOUT_TEST_IDS } from './Checkout.testIds';
